more testing

+++ Plutus/src/Serialize/SceneLoader.cpp
@@ -114,7 +114,7 @@ namespace Plutus
             auto docObj = doc.GetJsonObject();
             jhelper.value = &docObj;
 
-            scene->mBGColor = jhelper.getInt("bg-color");
+            scene->mBGColor = (int)jhelper.getInt("bg-color");
 
             if (doc.HasMember("fonts") && doc["fonts"].IsArray())
             {
